pyproject.toml
[project]
name = "axedit"
version = "0.1.0.dev"
dependencies = ["pygame-ce", "click"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]
axe = "axedit:cli.run"
axedit = "axedit:cli.run"
